Autonomous should work in theory. !HAS NOT BEEN TESTED!
Added and finished the class BezierCurve except for documentation.

Added and finished the class BezierCurveRoute except for documentation.

Added and finished the class CheckPoint except for documentation.

Added and finished the class CheckPoints except for documentation.

Added and finished the class LineSegment except for documentation.

Added and finished the class Point except for documentation.

Added the Class test which should theoreticaly work.